+++ src/main/kotlin/com/zepben/evolve/database/sqlite/common/BaseDatabaseReader.kt
@@ -9,16 +9,13 @@
 package com.zepben.evolve.database.sqlite.common
 
 import com.zepben.evolve.database.sqlite.tables.MissingTableConfigException
-import com.zepben.evolve.database.sqlite.upgrade.EwbDatabaseType
-import com.zepben.evolve.database.sqlite.upgrade.UpgradeRunner
+import com.zepben.evolve.database.sqlite.tables.TableVersion
 import com.zepben.evolve.services.common.BaseService
 import com.zepben.evolve.services.common.extensions.typeNameAndMRID
 import com.zepben.evolve.services.common.meta.MetadataCollection
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.nio.file.Paths
 import java.sql.Connection
-import java.sql.SQLException
 
 /**
  * A base class for reading objects from one of our databases.
@@ -27,7 +24,6 @@ import java.sql.SQLException
  * @property createMetadataReader Callback to create the reader for the [MetadataCollection] included in this database using the provided connection.
  * @property createServiceReader Callback to create the reader for the [BaseService] supported by this database using the provided connection.
  * @property service The [BaseService] that will be populated by the [BaseServiceReader]. Used for post-processing.
- * @property upgradeRunner The [UpgradeRunner] used to ensure this database is on the correct schema version.
  *
  * @property logger The [Logger] to use for this reader.
  * @property databaseDescriptor The JDBC descriptor for connecting to the database.
@@ -37,15 +33,16 @@ abstract class BaseDatabaseReader(
     private val createMetadataReader: (Connection) -> MetadataCollectionReader,
     private val createServiceReader: (Connection) -> BaseServiceReader,
     protected open val service: BaseService,
-    private val upgradeRunner: UpgradeRunner,
-    private val databaseType: EwbDatabaseType
+    private val createConnection: (String) -> Connection,
+    private val tableVersion: TableVersion
 ) {
 
     protected val logger: Logger = LoggerFactory.getLogger(javaClass)
 
     private val databaseDescriptor: String = "jdbc:sqlite:$databaseFile"
-    private lateinit var loadConnection: Connection
     private var hasBeenUsed: Boolean = false
+
+    private val supportedVersion = tableVersion.SUPPORTED_VERSION
 
     /**
      * Customisable function for performing actions after the database has been loaded.
@@ -73,43 +70,57 @@ abstract class BaseDatabaseReader(
         }
         hasBeenUsed = true
 
-        val databaseVersion = preLoad()
-        if (databaseVersion == null) {
-            closeConnection()
-            return false
-        }
-
-        logger.info("Loading from database version v$databaseVersion")
-        return try {
-            createMetadataReader(loadConnection).load() and
-                createServiceReader(loadConnection).load() and
-                postLoad()
-        } catch (e: MissingTableConfigException) {
-            logger.error("Unable to load database: " + e.message)
-            false
-        } finally {
-            closeConnection()
-        }
+        return preLoad()?.use { connection ->
+            try {
+                // NOTE: We only want to post process if we correctly read from the database.
+                loadFromReaders(connection)
+                    && postLoad()
+            } catch (e: MissingTableConfigException) {
+                logger.error("Unable to load database: " + e.message)
+                false
+            }
+        } ?: false
     }
 
-    private fun preLoad(): Int? =
+    private fun preLoad(): Connection? =
         try {
-            upgradeRunner.connectAndUpgrade(databaseDescriptor, Paths.get(databaseFile), databaseType)
-                .also { loadConnection = it.connection }
-                .version
-        } catch (e: UpgradeRunner.UpgradeException) {
+            val connection = createConnection(databaseDescriptor)
+            if (connection.checkVersion())
+                connection
+            else {
+                connection.close()
+                null
+            }
+        } catch (e: Exception) {
             logger.error("Failed to connect to the database for reading: " + e.message, e)
-            closeConnection()
             null
         }
 
-    private fun closeConnection() {
-        try {
-            if (::loadConnection.isInitialized)
-                loadConnection.close()
-        } catch (e: SQLException) {
-            logger.error("Failed to close connection to database: " + e.message)
+    private fun loadFromReaders(connection: Connection) =
+        createMetadataReader(connection).load() and
+            createServiceReader(connection).load()
+
+    private fun Connection.checkVersion(): Boolean =
+        createStatement().use { statement ->
+            val version = tableVersion.getVersion(statement)
+            if (version == supportedVersion) {
+                logger.info("Loading from database version v$version")
+                true
+            } else {
+                logger.error(formatVersionError(version))
+                false
+            }
         }
-    }
+
+    private fun formatVersionError(version: Int?): String =
+        when {
+            version == null -> "Failed to read the version number form the selected database. Are you sure it is a EWB database?"
+            version < supportedVersion -> unexpectedVersion(version, "Consider using the UpgradeRunner if you wish to support this database.")
+            else -> unexpectedVersion(version, "You need to use a newer version of the SDK to load this database.")
+        }
+
+
+    private fun unexpectedVersion(version: Int?, action: String) =
+        "Unable to load from database $databaseFile [found v$version, expected v$supportedVersion]. $action"
 
 }