+++ src/main/kotlin/com/zepben/evolve/database/sqlite/tables/TableVersion.kt
@@ -8,8 +8,18 @@
 
 package com.zepben.evolve.database.sqlite.tables
 
+import com.zepben.evolve.database.sqlite.extensions.executeConfiguredQuery
 import com.zepben.evolve.database.sqlite.tables.Column.Nullable.NOT_NULL
+import java.sql.SQLException
+import java.sql.Statement
 
+/**
+ * Code representation of the `version` table.
+ *
+ * @property SUPPORTED_VERSION The supported schema version.
+ *
+ * @property VERSION Column definition.
+ */
 @Suppress("PropertyName")
 class TableVersion : SqliteTable() {
 
@@ -18,5 +28,17 @@ class TableVersion : SqliteTable() {
     val VERSION: Column = Column(++columnIndex, "version", "TEXT", NOT_NULL)
 
     override val name: String = "version"
+
+    /**
+     * Helper function to read the version from the database.
+     */
+    @Throws(SQLException::class)
+    fun getVersion(statement: Statement): Int? =
+        runCatching {
+            statement.executeConfiguredQuery(selectSql).use { results ->
+                results.next() // todo needed? it wasn't here originally so how did that work?
+                results.getInt(VERSION.queryIndex)
+            }
+        }.getOrNull()
 
 }