+++ src/test/kotlin/com/zepben/evolve/database/sqlite/diagram/DiagramDatabaseReaderTest.kt
@@ -10,8 +10,6 @@ package com.zepben.evolve.database.sqlite.diagram
 
 import com.zepben.evolve.database.sqlite.common.MetadataCollectionReader
 import com.zepben.evolve.database.sqlite.tables.TableVersion
-import com.zepben.evolve.database.sqlite.upgrade.EwbDatabaseType
-import com.zepben.evolve.database.sqlite.upgrade.UpgradeRunner
 import com.zepben.evolve.services.diagram.DiagramService
 import com.zepben.testutils.junit.SystemLogExtension
 import io.mockk.every
@@ -21,8 +19,8 @@ import io.mockk.verifySequence
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
-import java.nio.file.Paths
 import java.sql.Connection
+import java.sql.Statement
 
 internal class DiagramDatabaseReaderTest {
 
@@ -36,12 +34,17 @@ internal class DiagramDatabaseReaderTest {
     private val metadataReader = mockk<MetadataCollectionReader>().also { every { it.load() } returns true }
     private val diagramServiceReader = mockk<DiagramServiceReader>().also { every { it.load() } returns true }
 
-    private val connection = mockk<Connection> { justRun { close() } }
-    private val connectionResult = mockk<UpgradeRunner.ConnectionResult>().also {
-        every { it.connection } returns connection
-        every { it.version } returns TableVersion().SUPPORTED_VERSION
+    private val statement = mockk<Statement> { justRun { close() } }
+    private val connection = mockk<Connection> {
+        every { createStatement() } returns statement
+        justRun { close() }
     }
-    private val upgradeRunner = mockk<UpgradeRunner>().also { every { it.connectAndUpgrade(any(), any(), any()) } returns connectionResult }
+    private val createConnection = mockk<(String) -> Connection>().also { every { it(any()) } returns connection }
+
+    private val tableVersion = mockk<TableVersion> {
+        every { getVersion(any()) } returns 1
+        every { SUPPORTED_VERSION } returns 1
+    }
 
     private val reader = DiagramDatabaseReader(
         databaseFile,
@@ -50,7 +53,8 @@ internal class DiagramDatabaseReaderTest {
         mockk(), // tables should not be used if we provide the rest of the parameters, so provide a mockk that will throw if used.
         { metadataReader },
         { diagramServiceReader },
-        upgradeRunner
+        createConnection,
+        tableVersion
     )
 
     //
@@ -62,7 +66,12 @@ internal class DiagramDatabaseReaderTest {
         assertThat("Should have loaded", reader.load())
 
         verifySequence {
-            upgradeRunner.connectAndUpgrade(match { it.contains(databaseFile) }, Paths.get(databaseFile), EwbDatabaseType.DIAGRAM)
+            tableVersion.SUPPORTED_VERSION
+            createConnection(match { it.contains(databaseFile) })
+            connection.createStatement()
+            tableVersion.getVersion(statement)
+            statement.close()
+
             metadataReader.load()
             diagramServiceReader.load()
             service.unresolvedReferences()